MONKEY-10: Make GNSS failure activity more informative
