+++ GPSTest/src/main/java/com/chesapeaketechnology/gnssmonkey/service/GpsMonkeyService.java
@@ -3,7 +3,8 @@ package com.chesapeaketechnology.gnssmonkey.service;
 import com.android.gpstest.Application;
 import com.android.gpstest.GpsTestListener;
 import com.android.gpstest.R;
-import com.chesapeaketechnology.gnssmonkey.FailureActivity;
+import com.android.gpstest.util.PreferenceUtils;
+import com.chesapeaketechnology.gnssmonkey.RawGnssFailureActivity;
 
 import android.Manifest;
 import android.app.Notification;
@@ -101,7 +102,14 @@ public class GpsMonkeyService extends Service {
                         firstGpsAcqTime = System.currentTimeMillis();
                     } else if (System.currentTimeMillis() > firstGpsAcqTime + TIME_TO_WAIT_FOR_GNSS_RAW_BEFORE_FAILURE) {
                         hasGnssRawFailureNagLaunched = true;
-                        startActivity(new Intent(GpsMonkeyService.this, FailureActivity.class));
+
+                        // The user may choose to continue using the app even without GNSS since
+                        // they do get some satellite status on this display. If that is the case,
+                        // they can choose not to be nagged about this every time they launch the app.
+                        boolean ignoreRawGnssFailure = PreferenceUtils.getBoolean(Application.get().getString(R.string.pref_key_ignore_raw_gnss_failure), false);
+                        if (!ignoreRawGnssFailure) {
+                            startActivity(new Intent(GpsMonkeyService.this, RawGnssFailureActivity.class));
+                        }
                     }
                 }
             }
@@ -233,7 +241,8 @@ public class GpsMonkeyService extends Service {
     }
 
     /**
-     * Running GPSMonkey as a foreground service allows GPSMonkey to stay active on API level 26+ devices.
+     * Running GPSMonkey as a foreground service allows GPSMonkey to stay active on API level 26+
+     * devices.
      * Depending on desired collection rates, could also consider migrating to a JobScheduler.
      */
     private void setForeground() {
