+++ GPSTest/src/main/res/layout/activity_failure.xml
@@ -1,11 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".FailureActivity">
+        xmlns:tools="http://schemas.android.com/tools"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        tools:context=".FailureActivity">
 
     <ImageView
         android:id="@+id/aboutLogo"
@@ -14,17 +13,27 @@
         android:src="@mipmap/ic_launcher_round" />
 
     <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center_horizontal"
-        android:text="@string/not_supported_title"
-        android:textSize="18sp"
-        android:textStyle="bold" />
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:gravity="center_horizontal"
+            android:text="@string/not_supported_title"
+            android:textSize="24sp"
+            android:textStyle="bold" />
 
     <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/not_supported_description"/>
+            android:id="@+id/failureDescriptionTextView"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:paddingLeft="6dp"
+            android:paddingRight="6dp"
+            android:text="@string/not_supported_description"
+            android:textSize="18sp" />
+
+    <CheckBox
+            android:id="@+id/failureRememberDecisionCheckBox"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="@string/remember_my_decision" />
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -32,11 +41,11 @@
         android:orientation="horizontal">
 
         <Button
-            android:id="@+id/failureButton"
+                android:id="@+id/failureOkButton"
             android:layout_width="0dp"
             android:layout_weight="1"
             android:layout_height="wrap_content"
-            android:text="@string/try_anyway"/>
+                android:text="@string/use_anyway" />
 
         <Button
             android:id="@+id/failureUninstallButton"
