+++ GPSTest/src/main/res/values/strings.xml
@@ -438,7 +438,8 @@
     <string name="share_geopackage_narrative">Share recorded data to others.</string>
     <string name="share_geopackage">Share Data</string>
     <string name="not_supported_title">GPS Monkey may not work.</string>
-    <string name="not_supported_description">GPS Monkey is not receiving raw measurements from the phone.</string>
-    <string name="try_anyway">Keep Trying...</string>
-    <string name="uninstall">Uninstall GPS Monkey</string>
+    <string name="not_supported_description">GPS Monkey is not receiving raw GNSS measurements from the phone, which are required for generating GeoPackage files with useful data. You will still see some satellite status in the app, but any GeoPackage files generated by GPS Monkey will only contain basic location information. For more information and a list of devices that support raw GNSS, consult Android\'s <a href="https://developer.android.com/guide/topics/sensors/gnss">Raw GNSS Measurements</a> guide.</string>
+    <string name="remember_my_decision">Don\'t ask me again</string>
+    <string name="use_anyway">Use anyway</string>
+    <string name="uninstall">Uninstall</string>
 </resources>
